+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2026-06-20
+
+### Added
+
+- Added Create haunting recipes for all soul-related items.
+- Added additional item tags for brick grouping.
+- Added Kilns to the `#supplementaries:bellows_tickable` tag (@Bobisnotap).
+- Kilns now output a comparator signal akin to furnaces based on the fullness of their slots (@Bobisnotap).
+- Kilns now detect soul fire under hopper blocks (@Bobisnotap).
+
 ## [1.1.0] - 2026-06-02
 
 ### Added